finished implementing reverse kinematic
