+++ launch/cam.launch
@@ -1,6 +1,6 @@
 <?xml version='1.0'?>
 <launch>
-  <node name="arm-cam" pkg="usb_cam" type="usb_cam_node" output="screen" >
+  <node name="arm_cam" pkg="usb_cam" type="usb_cam_node" output="screen" >
     <param name="video_device" value="/dev/video2" />
     <param name="framerate" value="30" />
     <param name="image_width" value="1920" />
@@ -11,7 +11,7 @@
     <param name="io_method" value="mmap"/>
   </node>
 
-  <node name="arm-cam-2" pkg="usb_cam" type="usb_cam_node" output="screen" >
+  <node name="arm_cam-2" pkg="usb_cam" type="usb_cam_node" output="screen" >
     <param name="video_device" value="/dev/video4" />
     <param name="framerate" value="30" />
     <param name="image_width" value="640" />
