--- launch/simulation.launch
@@ -1,63 +1,0 @@
-<?xml version='1.0'?>
-<launch>
-	<!-- Launch robot bringup -->
-	<!--
-	<include file="$(find zeus_arm)/launch/bringup.launch" />
-	-->
-
-	<!-- Launch Gazebo with an empty world -->
-	<include file="$(find gazebo_ros)/launch/empty_world.launch">
-	    <arg name="gui" value="true" />
-	    <arg name="use_sim_time" value="true"/>
-	    <arg name="debug" value="false" />
-	    <arg name="paused" value="true"/>
-	</include>
-
-	<!-- Load the description for the robot -->
-	<param name="robot_description" command="cat '$(find zeus_arm)/urdf/robot.urdf'" />
-
-	<!-- Spawn the robot in Gazebo -->
-	<node name="spawn_robot_urdf" pkg="gazebo_ros" type="spawn_model" output="screen"
-		args="-urdf -param robot_description -model zeus_arm -robot_namespace /zeus_arm 
-		              -x 0 
-                      -y 0 
-                      -z 0.5
-                      -J joint_1 0
-                      -J joint_2 0
-                      -J joint_3 0
-                      -J joint_4 0" />
-
-	<!-- Load controller configuration file -->
-	<rosparam file="$(find zeus_arm)/config/joint_velocity_controllers.yaml" command="load"/>
-
-	<!-- Load controller spawner from ROS control -->
-
-	<node  name="controller_spawner" pkg="controller_manager" type="spawner" respawn="false" output="screen" 
-	 ns="/zeus_arm" args="joint_1_velocity_controller 
-			     		  joint_2_velocity_controller 
-                          joint_3_velocity_controller 
-                          joint_4_velocity_controller 
-                          joint_state_controller"/>
-
-    <!-- Launch joint state publisher from ROS control -->
-	<node name="robot_state_publisher" pkg="robot_state_publisher" type="robot_state_publisher" respawn="false" output="screen">
-		<remap from="/joint_states" to ="/zeus_arm/joint_states" />
-	</node>
-
-
-	<!-- Launch teleop script -->
-	<!--
-	<node name="teleop_node" pkg="zeus_arm" type="pos_teleop.py" output="screen" />
-	-->
-
-	<!-- Launch arm node -->
-	<!--
-	<node name="arm_node" pkg="zeus_arm" type="arm_node.py" output="screen" />
-	-->
-
-    <!-- Initialize the robot and unpause gazebo -->
-	<arg name="start_delay_seconds" default="5"/>
-	<node name="init_robot" pkg="zeus_arm" type="init_robot.py" output="screen" respawn="false" launch-prefix="bash -c 'sleep $(arg start_delay_seconds); $0 $@'"/>
-
-
-</launch>
